feat: add ImageHeaderCollectionReusableView for main image
